Move ParcelableDeferredBoolean to a legal Java package
